update bug & feature superadmin admin user

+++ database/migrations/2025_07_06_042746_jenis_pekerjaans.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('nama_pekerjaan');
             $table->string('satuan');
-            $table->integer('volume')->default(0); // ✅ Tambahan kolom volume
-            $table->integer('bobot')->default(0); // ✅ Tambahan kolom bobot 1-100
+            $table->integer('volume')->default(0);
+            $table->double('bobot')->default(0);
             $table->string('pemberi_pekerjaan')->nullable();
             $table->foreignId('tim_id')->nullable()->constrained('teams')->onDelete('set null');
             $table->timestamps();
